+++ pkg/timer_event/setup.go
@@ -12,11 +12,12 @@ import (
 
 // UserManagementTimerService handles background times for user management (cleanup for example).
 type UserManagementTimerService struct {
-	globalDBService      *globaldb.GlobalDBService
-	userDBService        *userdb.UserDBService
-	clients              *models.APIClients
-	TimerEventFrequency  int64 // how often the timer event should be performed (only from one instance of the service) - seconds
-	CleanUpTimeThreshold int64 // if user account not verified, remove user after this many seconds
+	globalDBService       *globaldb.GlobalDBService
+	userDBService         *userdb.UserDBService
+	clients               *models.APIClients
+	TimerEventFrequency   int64 // how often the timer event should be performed (only from one instance of the service) - seconds
+	CleanUpTimeThreshold  int64 // if user account not verified, remove user after this many seconds
+	ReminderTimeThreshold int64 // if user account not verified, send a reminder email to the user after this many seconds
 }
 
 func NewUserManagmentTimerService(
@@ -25,13 +26,15 @@ func NewUserManagmentTimerService(
 	userDBService *userdb.UserDBService,
 	clients *models.APIClients,
 	cleanUpTimeThreshold int64,
+	reminderTimeThreshold int64,
 ) *UserManagementTimerService {
 	return &UserManagementTimerService{
-		globalDBService:      globalDBService,
-		userDBService:        userDBService,
-		TimerEventFrequency:  frequency,
-		clients:              clients,
-		CleanUpTimeThreshold: cleanUpTimeThreshold,
+		globalDBService:       globalDBService,
+		userDBService:         userDBService,
+		TimerEventFrequency:   frequency,
+		clients:               clients,
+		CleanUpTimeThreshold:  cleanUpTimeThreshold,
+		ReminderTimeThreshold: reminderTimeThreshold,
 	}
 }
 
@@ -45,7 +48,7 @@ func (s *UserManagementTimerService) startTimerThread(ctx context.Context, timeC
 		select {
 		case <-time.After(time.Duration(timeCheckInterval) * time.Second):
 			go s.CleanUpUnverifiedUsers()
-
+			go s.ReminderToConfirmAccount()
 		case <-ctx.Done():
 			return
 		}